more nits
