+++ plugins/source/tailscale/resources/services/acl/acls.go
@@ -1,6 +1,8 @@
 package acl
 
 import (
+	"context"
+
 	"github.com/cloudquery/cloudquery/plugins/source/tailscale/client"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
@@ -49,4 +51,15 @@ func Acls() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchAcls(ctx context.Context, meta schema.ClientMeta, _ *schema.Resource, res chan<- any) error {
+	c := meta.(*client.Client)
+	result, err := c.TailscaleClient.ACL(ctx)
+	if err != nil {
+		return err
+	}
+
+	res <- result
+	return nil
 }
