--- plugins/source/tailscale/resources/services/acl/acls_fetch.go
@@ -1,19 +1,0 @@
-package acl
-
-import (
-	"context"
-
-	"github.com/cloudquery/cloudquery/plugins/source/tailscale/client"
-	"github.com/cloudquery/plugin-sdk/schema"
-)
-
-func fetchAcls(ctx context.Context, meta schema.ClientMeta, _ *schema.Resource, res chan<- any) error {
-	c := meta.(*client.Client)
-	result, err := c.TailscaleClient.ACL(ctx)
-	if err != nil {
-		return err
-	}
-
-	res <- result
-	return nil
-}
