--- plugins/source/tailscale/resources/services/device/devices_fetch.go
@@ -1,20 +1,0 @@
-package device
-
-import (
-	"context"
-
-	"github.com/cloudquery/cloudquery/plugins/source/tailscale/client"
-	"github.com/cloudquery/plugin-sdk/schema"
-)
-
-func fetchDevices(ctx context.Context, meta schema.ClientMeta, _ *schema.Resource, res chan<- any) error {
-	c := meta.(*client.Client)
-
-	result, err := c.TailscaleClient.Devices(ctx)
-	if err != nil {
-		return err
-	}
-
-	res <- result
-	return nil
-}
