+++ plugins/source/tailscale/resources/services/device/devices_mock_test.go
@@ -32,7 +32,24 @@ func createDevices(mux *httprouter.Router) error {
 		}
 	})
 
-	return createRoutes(mux)
+	var route tailscale.DeviceRoutes
+	if err := faker.FakeObject(&route); err != nil {
+		return err
+	}
+
+	mux.GET("/api/v2/device/:device/routes", func(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+		b, err := json.Marshal(route)
+		if err != nil {
+			http.Error(w, "unable to marshal response: "+err.Error(), http.StatusBadRequest)
+			return
+		}
+		if _, err := w.Write(b); err != nil {
+			http.Error(w, "failed to write", http.StatusBadRequest)
+			return
+		}
+	})
+
+	return nil
 }
 
 func TestDevices(t *testing.T) {
