+++ plugins/source/tailscale/resources/services/key/keys.go
@@ -1,6 +1,8 @@
 package key
 
 import (
+	"context"
+
 	"github.com/cloudquery/cloudquery/plugins/source/tailscale/client"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
@@ -25,4 +27,28 @@ func Keys() *schema.Table {
 			},
 		},
 	}
+}
+
+func fetchKeys(ctx context.Context, meta schema.ClientMeta, _ *schema.Resource, res chan<- any) error {
+	c := meta.(*client.Client)
+
+	result, err := c.TailscaleClient.Keys(ctx)
+	if err != nil {
+		return err
+	}
+
+	res <- result
+	return nil
+}
+
+func getKey(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource) error {
+	c := meta.(*client.Client)
+	key := resource.Item.(tailscale.Key)
+	result, err := c.TailscaleClient.GetKey(ctx, key.ID)
+	if err != nil {
+		return err
+	}
+
+	resource.SetItem(result)
+	return nil
 }
